+++ src/actions/index.js
@@ -7,6 +7,7 @@ import * as bidActions from './bidActions'
 import * as transactionActions from './transactionActions'
 import * as validationActions from './validationActions'
 import * as analyticsActions from './analyticsActions'
+import * as campaignActions from './campaignActions'
 
 // keep it for class components that use connect and actions prop
 // without needing to change the code
@@ -20,6 +21,7 @@ export default {
 	...transactionActions,
 	...validationActions,
 	...analyticsActions,
+	...campaignActions,
 }
 
 // used for minimal deps in hook components
@@ -34,3 +36,4 @@ export * from './bidActions'
 export * from './transactionActions'
 export * from './validationActions'
 export * from './analyticsActions'
+export * from './campaignActions'
